refactor: 한장의 카드를 Deck에서 뽑아주는 기능 리팩토링
