+++ src/main/java/blackjack/domain/Deck.java
@@ -1,13 +1,17 @@
 package blackjack.domain;
 
-import blackjack.constant.Denomination;
-import blackjack.constant.Type;
+import blackjack.enums.Denomination;
+import blackjack.enums.Type;
 import lombok.Getter;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Random;
 
 public class Deck {
+
+    private static final Random RANDOM = new Random();
+    private static final String ALERT_NO_CARD_LEFT = "사용 가능한 카드를 모두 소진하였습니다.";
 
     @Getter
     private final List<Card> deck = new LinkedList<>();
